Start actually pronouncing letters
We start with possibly the simplest known Elvish word (though not the
shortest by number of syllables, I assume): Eä, the World that Is.

The pronunciation seems already arguable: Tolkien Gateway gives Eä as
[ˈe.a] – e rather than ɛ, but a rather than ɑ. I half disagree and
produce ˈeɑ for now.

The test setup is partially based on tolkien-stress again, and (at least
for pronounceWord()) intends to make it fairly easy to add further test
cases. I don’t think we can avoid actually spelling out the IPA, but we
can use the same no|TA|tion as in tolkien-stress to encode syllable
breaks and stress in a test case without having to spell them out each
time; this keeps the test independent of the real analyseWord.