Pronounce “Eru”
Appendix E says R is always trilled, and for now we always give a voiced
alveolar trill. Personally, I sometimes say it as a simpler alveolar tap
(ɾ), but so far I haven’t figured out my own rules for doing that. I
think we can consider them the same phoneme for now (and aim for
phonemic, not phonetic, notation).